Migrate Playground JavaScript to Stimulus and improve mobile styles
